Remove undeterministic Test for non-existant urls
The test would fail on linux machines unexpectantly. Initial ideas
is that it relies on an os function around the use of Sockets to
connect to URLs. In these instances, rather than error it would connect
and return an empty document.

As other tests cover connection failure use cases, as well as cases with
a valid URL but no html/data/product data, the decision was to remove the test
instead of @Ignoring it.